+++ src/api/java/com/minecolonies/api/entity/ai/combat/threat/IThreatTableEntity.java
@@ -1,4 +1,4 @@
-package com.minecolonies.api.entity.combat.threat;
+package com.minecolonies.api.entity.ai.combat.threat;
 
 /**
  * Entities implement this for the necessary hooks
