+++ src/api/java/com/minecolonies/api/entity/pathfinding/PathingStuckHandler.java
@@ -1,7 +1,7 @@
 package com.minecolonies.api.entity.pathfinding;
 
 import com.ldtteam.structurize.util.BlockUtils;
-import com.minecolonies.api.entity.ai.citizen.builder.IBuilderUndestroyable;
+import com.minecolonies.api.entity.ai.workers.builder.IBuilderUndestroyable;
 import com.minecolonies.api.items.ModTags;
 import com.minecolonies.api.util.BlockPosUtil;
 import com.minecolonies.api.util.DamageSourceKeys;
