+++ src/main/java/com/minecolonies/core/entity/ai/combat/AttackMoveAI.java
@@ -3,9 +3,9 @@ package com.minecolonies.core.entity.ai.combat;
 import com.minecolonies.api.entity.ai.statemachine.states.IState;
 import com.minecolonies.api.entity.ai.statemachine.tickratestatemachine.ITickRateStateMachine;
 import com.minecolonies.api.entity.ai.statemachine.tickratestatemachine.TickingTransition;
-import com.minecolonies.api.entity.combat.CombatAIStates;
-import com.minecolonies.api.entity.combat.threat.IThreatTableEntity;
-import com.minecolonies.api.entity.combat.threat.ThreatTableEntry;
+import com.minecolonies.api.entity.ai.combat.CombatAIStates;
+import com.minecolonies.api.entity.ai.combat.threat.IThreatTableEntity;
+import com.minecolonies.api.entity.ai.combat.threat.ThreatTableEntry;
 import com.minecolonies.api.entity.pathfinding.AbstractAdvancedPathNavigate;
 import com.minecolonies.api.entity.pathfinding.PathResult;
 import com.minecolonies.api.util.DamageSourceKeys;
