+++ src/main/java/com/minecolonies/core/entity/pathfinding/EntityCitizenWalkToProxy.java
@@ -1,7 +1,7 @@
 package com.minecolonies.core.entity.pathfinding;
 
 import com.minecolonies.api.colony.buildings.IBuilding;
-import com.minecolonies.api.entity.ai.pathfinding.AbstractWalkToProxy;
+import com.minecolonies.api.entity.pathfinding.proxy.AbstractWalkToProxy;
 import com.minecolonies.api.entity.citizen.AbstractEntityCitizen;
 import com.minecolonies.api.util.BlockPosUtil;
 import com.minecolonies.api.util.EntityUtils;
