+++ src/main/java/com/minecolonies/core/entity/pathfinding/GeneralEntityWalkToProxy.java
@@ -2,7 +2,7 @@ package com.minecolonies.core.entity.pathfinding;
 
 import com.minecolonies.api.colony.IColony;
 import com.minecolonies.api.colony.IColonyManager;
-import com.minecolonies.api.entity.ai.pathfinding.AbstractWalkToProxy;
+import com.minecolonies.api.entity.pathfinding.proxy.AbstractWalkToProxy;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.core.BlockPos;
