+++ src/main/java/com/minecolonies/core/entity/visitor/VisitorCitizen.java
@@ -6,7 +6,7 @@ import com.minecolonies.api.colony.permissions.Action;
 import com.minecolonies.api.colony.requestsystem.StandardFactoryController;
 import com.minecolonies.api.colony.requestsystem.location.ILocation;
 import com.minecolonies.api.entity.CustomGoalSelector;
-import com.minecolonies.api.entity.ai.pathfinding.IWalkToProxy;
+import com.minecolonies.api.entity.pathfinding.proxy.IWalkToProxy;
 import com.minecolonies.api.entity.citizen.AbstractEntityCitizen;
 import com.minecolonies.api.entity.citizen.citizenhandlers.*;
 import com.minecolonies.api.inventory.InventoryCitizen;
@@ -23,8 +23,6 @@ import com.minecolonies.core.entity.ai.minimal.LookAtEntityGoal;
 import com.minecolonies.core.entity.ai.visitor.EntityAIVisitor;
 import com.minecolonies.core.entity.citizen.citizenhandlers.*;
 import com.minecolonies.core.entity.citizen.EntityCitizen;
-import com.minecolonies.core.entity.entity.citizen.citizenhandlers.*;
-import com.minecolonies.core.entity.entity.citizenhandlers.*;
 import com.minecolonies.core.entity.pathfinding.EntityCitizenWalkToProxy;
 import com.minecolonies.core.entity.pathfinding.MovementHandler;
 import com.minecolonies.core.network.messages.client.ItemParticleEffectMessage;
